contract data updated

+++ src/main/java/com/synergizglobal/pmis/IMPLdao/ContractDaoImpl.java
@@ -538,35 +538,38 @@ public class ContractDaoImpl implements ContractDao {
 			    c = stmt.executeBatch();
 				if(stmt != null){stmt.close();}
 				
-				PreparedStatement stmtRevision = null;
-				
-				String deleteQryRevision = "DELETE from contract_revisions where contract_id_fk = ?";		 
-				stmtRevision = con.prepareStatement(deleteQryRevision);
-				stmtRevision.setString(1,contract.getContract_id()); 
-				stmtRevision.executeUpdate();
-				if(stmtRevision != null){stmtRevision.close();}					
-				
-				
-				PreparedStatement preparedStmtRevisions = null;
-				String queryTdcRevisions = " insert into contract_revisions (contract_id_fk,revision_no,revision_estimated_cost,revision_planned_date_of_award,revision_planned_date_of_completion)"
-			               + " values (?,?,?,?,?)";				
-				
-				preparedStmtRevisions = con.prepareStatement(queryTdcRevisions);
-
-				if(!StringUtils.isEmpty(contract) && !StringUtils.isEmpty(contract.getRevisionno())) 
+				if(!StringUtils.isEmpty(contract.getRevision_estimated_cost()) && contract.getRevision_estimated_cost().length>0 && contract.getRevision_estimated_cost()!=null)
 				{
-
-					for (int i = 0; i < contract.getRevisionno().length; i++) 
+					PreparedStatement stmtRevision = null;
+					
+					String deleteQryRevision = "DELETE from contract_revisions where contract_id_fk = ?";		 
+					stmtRevision = con.prepareStatement(deleteQryRevision);
+					stmtRevision.setString(1,contract.getContract_id()); 
+					stmtRevision.executeUpdate();
+					if(stmtRevision != null){stmtRevision.close();}					
+					
+					
+					PreparedStatement preparedStmtRevisions = null;
+					String queryTdcRevisions = " insert into contract_revisions (contract_id_fk,revision_no,revision_estimated_cost,revision_planned_date_of_award,revision_planned_date_of_completion)"
+				               + " values (?,?,?,?,?)";				
+					
+					preparedStmtRevisions = con.prepareStatement(queryTdcRevisions);
+	
+					if(!StringUtils.isEmpty(contract) && !StringUtils.isEmpty(contract.getRevisionno())) 
 					{
-						preparedStmtRevisions.setString(1, contract_id);
-						preparedStmtRevisions.setString(2, contract.getRevisionno()[i]);
-						preparedStmtRevisions.setString(3, contract.getRevision_estimated_cost()[i]);
-						preparedStmtRevisions.setString(4, contract.getRevision_planned_date_of_award()[i]);
-						preparedStmtRevisions.setString(5, contract.getRevision_planned_date_of_completion()[i]);						
-						preparedStmtRevisions.execute();
+	
+						for (int i = 0; i < contract.getRevisionno().length; i++) 
+						{
+							preparedStmtRevisions.setString(1, contract_id);
+							preparedStmtRevisions.setString(2, contract.getRevisionno()[i]);
+							preparedStmtRevisions.setString(3, contract.getRevision_estimated_cost()[i]);
+							preparedStmtRevisions.setString(4, contract.getRevision_planned_date_of_award()[i]);
+							preparedStmtRevisions.setString(5, contract.getRevision_planned_date_of_completion()[i]);						
+							preparedStmtRevisions.execute();
+						}
 					}
+					if(preparedStmtRevisions != null){preparedStmtRevisions.close();}	
 				}
-				if(preparedStmtRevisions != null){preparedStmtRevisions.close();}				
 				
 				
 				String Insurence_qry = "INSERT into  insurance (insurance_type_fk,issuing_agency,agency_address,"
@@ -1891,35 +1894,39 @@ public class ContractDaoImpl implements ContractDao {
 					int[] updateCount = updateStmt.executeBatch();
 					if(stmt != null || updateStmt != null ){stmt.close();updateStmt.close();}
 					
-					PreparedStatement stmtRevision = null;
-					
-					String deleteQryRevision = "DELETE from contract_revisions where contract_id_fk = ?";		 
-					stmtRevision = con.prepareStatement(deleteQryRevision);
-					stmtRevision.setString(1,contract.getContract_id()); 
-					stmtRevision.executeUpdate();
-					if(stmtRevision != null){stmtRevision.close();}					
-					
-					
-					PreparedStatement preparedStmtRevisions = null;
-					String queryTdcRevisions = " insert into contract_revisions (contract_id_fk,revision_no,revision_estimated_cost,revision_planned_date_of_award,revision_planned_date_of_completion)"
-				               + " values (?,?,?,?,?)";				
-					
-					preparedStmtRevisions = con.prepareStatement(queryTdcRevisions);
-
-					if(!StringUtils.isEmpty(contract) && !StringUtils.isEmpty(contract.getRevisionno())) 
+					if(!StringUtils.isEmpty(contract.getRevision_estimated_cost()) && contract.getRevision_estimated_cost().length>0 && contract.getRevision_estimated_cost()!=null)
 					{
-
-						for (int i = 0; i < contract.getRevisionno().length; i++) 
+					
+						PreparedStatement stmtRevision = null;
+						
+						String deleteQryRevision = "DELETE from contract_revisions where contract_id_fk = ?";		 
+						stmtRevision = con.prepareStatement(deleteQryRevision);
+						stmtRevision.setString(1,contract.getContract_id()); 
+						stmtRevision.executeUpdate();
+						if(stmtRevision != null){stmtRevision.close();}					
+						
+						
+						PreparedStatement preparedStmtRevisions = null;
+						String queryTdcRevisions = " insert into contract_revisions (contract_id_fk,revision_no,revision_estimated_cost,revision_planned_date_of_award,revision_planned_date_of_completion)"
+					               + " values (?,?,?,?,?)";				
+						
+						preparedStmtRevisions = con.prepareStatement(queryTdcRevisions);
+	
+						if(!StringUtils.isEmpty(contract) && !StringUtils.isEmpty(contract.getRevisionno())) 
 						{
-							preparedStmtRevisions.setString(1, contract.getContract_id());
-							preparedStmtRevisions.setString(2, contract.getRevisionno()[i]);
-							preparedStmtRevisions.setString(3, contract.getRevision_estimated_cost()[i]);
-							preparedStmtRevisions.setString(4, contract.getRevision_planned_date_of_award()[i]);
-							preparedStmtRevisions.setString(5, contract.getRevision_planned_date_of_completion()[i]);						
-							preparedStmtRevisions.execute();
+	
+							for (int i = 0; i < contract.getRevisionno().length; i++) 
+							{
+								preparedStmtRevisions.setString(1, contract.getContract_id());
+								preparedStmtRevisions.setString(2, contract.getRevisionno()[i]);
+								preparedStmtRevisions.setString(3, contract.getRevision_estimated_cost()[i]);
+								preparedStmtRevisions.setString(4, contract.getRevision_planned_date_of_award()[i]);
+								preparedStmtRevisions.setString(5, contract.getRevision_planned_date_of_completion()[i]);						
+								preparedStmtRevisions.execute();
+							}
 						}
+						if(preparedStmtRevisions != null){preparedStmtRevisions.close();}	
 					}
-					if(preparedStmtRevisions != null){preparedStmtRevisions.close();}					
 					
 				
 					deleteQry = "DELETE from contract_revision where contract_id_fk = ?";
